Red Left was the one I was primarily changing
DEGREES_PER_COUNT works by taking a full circle or 360 degrees and then dividing that by the encoder count to get how many degrees of movement per encoder count.

INCHES_PER_DEGREE works by finding the total circumference of the arc of the arm at the end point, and then to find how many inches a degree of movement is it takes how much circumference one degree is

the last part just combines it and then gives how much the arm moves in inches for the encoder part. I was lowkey tripping earlier, idk why I did it this way instead of just taking the motor counts and the gear reduction so like 3420 encoder counts and then finding the circumference of the full circle so just 2 times pi times 17 and then dividing the circumference by the encoder counts but I guess it works either way